+++ src/providers/hive/auth.js
@@ -303,6 +303,7 @@ export const verifyPinCode = async (data) => {
     }
 
     if (result.length > 0) {
+      //TOOD: handle refresh token failure
       await refreshSCToken(userData, get(data, 'pinCode'));
     }
     return true;
@@ -314,19 +315,27 @@ export const verifyPinCode = async (data) => {
 
 export const refreshSCToken = async (userData, pinCode) => {
   const scAccount = await getSCAccount(userData.username);
-  const now = new Date();
-  const expireDate = new Date(scAccount.expireDate);
+  const now = new Date().getTime();
+  const expireDate = new Date(scAccount.expireDate).getTime();
 
-  const newSCAccountData = await getSCAccessToken(scAccount.refreshToken);
+  try {
+    const newSCAccountData = await getSCAccessToken(scAccount.refreshToken);
 
-  await setSCAccount(newSCAccountData);
-  const accessToken = newSCAccountData.access_token;
-  const encryptedAccessToken = encryptKey(accessToken, pinCode);
-  await updateUserData({
-    ...userData,
-    accessToken: encryptedAccessToken,
-  });
-  return encryptedAccessToken;
+    await setSCAccount(newSCAccountData);
+    const accessToken = newSCAccountData.access_token;
+    const encryptedAccessToken = encryptKey(accessToken, pinCode);
+    await updateUserData({
+      ...userData,
+      accessToken: encryptedAccessToken,
+    });
+    return encryptedAccessToken;
+  } catch (error) {
+    if (now > expireDate) {
+      throw error;
+    } else {
+      console.warn('token failed to refresh but current token is still valid');
+    }
+  }
 };
 
 export const switchAccount = (username) =>
